requirements.txt
# Core dependencies for AI Security Testing Application
PyYAML>=6.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0

# Optional dependencies for advanced features
transformers>=4.35.0
torch>=2.0.0
scikit-learn>=1.3.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Security and validation
cryptography>=41.0.0
jsonschema>=4.19.0
